# docs/swagger.yaml
basePath: /
definitions:
...
  model.Chapter:
    properties:
      course:
        $ref: '#/definitions/model.Course'
      courseId:
        type: integer
      id:
        type: integer
      link:
        type: string
      name:
        type: string
    type: object
  model.Course:
    properties:
      codeCourse:
        type: string
      id:
        type: integer
      lessons:
        type: integer
      name:
        type: string
      subject:
        $ref: '#/definitions/model.Subject'
      subjectId:
        type: integer
      teacher:
        $ref: '#/definitions/model.Teacher'
      teacherId:
        type: integer
    type: object
...
  model.Exercise:
    properties:
      course:
        $ref: '#/definitions/model.Course'
      courseId:
        type: integer
      id:
        type: integer
      link:
        type: string
      name:
        type: string
    type: object
...
  model.Subject:
    properties:
      codeSubject:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  model.Teacher:
    properties:
      email:
        type: string
      id:
        type: integer
      name:
        type: string
      pass:
        type: string
    type: object
host: localhost:3000
info:
  contact: {}
  description: Api canvas project
  title: Canvas Project
  version: "2.0"
paths:
...
  /teacher/courses/{id}:
    get:
      consumes:
      - application/json
      description: Course of teacher by id
      parameters:
      - description: Course id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with course
  /teacher/courses/{id}/attendance/{attendance_id}:
    put:
      consumes:
      - application/json
      description: Teacher update attendance status student
      parameters:
      - description: Course id
        in: path
        name: id
        required: true
        type: string
      - description: Attendance status id
        in: path
        name: attendance_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with attendance status
  /teacher/courses/{id}/chapter:
    get:
      consumes:
      - application/json
      description: Teacher get all chapter of course
      parameters:
      - description: Course Id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with chapter
  /teacher/courses/{id}/chapter/{chapter_id}:
    get:
      consumes:
      - application/json
      description: Teacher get chapter id
      parameters:
      - description: Course id
        in: path
        name: id
        required: true
        type: string
      - description: Chapter id
        in: path
        name: chapter_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with chapter
  /teacher/courses/{id}/chapter/create:
    post:
      consumes:
      - application/json
      description: Teacher create chapter for course
      parameters:
      - description: Course id
        in: path
        name: id
        required: true
        type: string
      - description: Chapter
        format: model.Chapter
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Chapter'
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with chapter
  /teacher/courses/{id}/chapter/delete:
    delete:
      consumes:
      - application/json
      description: Teacher delete chapter
      parameters:
      - description: Course id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with chapter
  /teacher/courses/{id}/chapter/update:
    put:
      consumes:
      - application/json
      description: Teacher update chapter
      parameters:
      - description: Course id
        in: path
        name: id
        required: true
        type: string
      - description: Chapter id
        in: path
        name: chapter_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with chapter
  /teacher/courses/{id}/exercise:
    get:
      consumes:
      - application/json
      description: Teacher get exercises of course
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with exercise
  /teacher/courses/{id}/exercise/{exercise_id}:
    get:
      consumes:
      - application/json
      description: Teacher get exercise by id
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      - description: execie id
        in: path
        name: exercise_id
        required: true
        type: string
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with exercise
...
  /teacher/courses/{id}/exercise/create:
    post:
      consumes:
      - application/json
      description: Teacher create exercise
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      - description: Exercise
        format: model.Exercise
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Exercise'
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with exercise
  /teacher/courses/{id}/exercise/delete:
    delete:
      consumes:
      - application/json
      description: Teacher delete exercise
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      - description: exercise need delete
        format: model.Exercise
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Exercise'
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with exercise
  /teacher/courses/{id}/exercise/update:
    put:
      consumes:
      - application/json
      description: Teacher update exercise of course
      parameters:
      - description: course id
        in: path
        name: id
        required: true
        type: string
      - description: exercise need update
        format: model.Exercise
        in: body
        name: req
        schema:
          $ref: '#/definitions/model.Exercise'
      produces:
      - application/json
      responses: {}
      tags:
      - Teacher with exercise
...
swagger: "2.0"